// sdk/node/src/proto/attune/v1/tag.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.6.1
//   protoc               unknown
// source: attune/v1/tag.proto

/* eslint-disable */

export const protobufPackage = "attune.v1";

export interface Tag {
  id: string;
  name: string;
  color: string;
  description: string;
  exclusiveScope?: string | undefined;
  usageCount: number;
  archived: boolean;
  createdBy: string;
  createdAt: string;
  updatedAt: string;
}

export interface ListTagsRequest {
  includeArchived: boolean;
}

export interface ListTagsResponse {
  tags: Tag[];
}

export interface CreateTagRequest {
  name: string;
  color?: string | undefined;
  description?: string | undefined;
  exclusiveScope?: string | undefined;
}

export interface UpdateTagRequest {
  id: string;
  name?: string | undefined;
  color?: string | undefined;
  description?: string | undefined;
  exclusiveScope?: string | undefined;
}

export interface ArchiveTagRequest {
  id: string;
}

export interface ArchiveTagResponse {
}

export interface AddFeedbackTagRequest {
  feedbackId: string;
  tagId?: string | undefined;
  tagName?: string | undefined;
  tagColor?: string | undefined;
}

export interface AddFeedbackTagResponse {
  tag?: Tag | undefined;
}

export interface RemoveFeedbackTagRequest {
  feedbackId: string;
  tagId: string;
}

export interface RemoveFeedbackTagResponse {
}

export interface BatchUpdateFeedbackTagsRequest {
  feedbackIds: string[];
  addTagIds: string[];
  removeTagIds: string[];
}

export interface BatchUpdateFeedbackTagsResponse {
  affected: number;
}

export interface FeedbackTagService {
  ListTags(request: ListTagsRequest): Promise<ListTagsResponse>;
  CreateTag(request: CreateTagRequest): Promise<Tag>;
  UpdateTag(request: UpdateTagRequest): Promise<Tag>;
  ArchiveTag(request: ArchiveTagRequest): Promise<ArchiveTagResponse>;
  AddFeedbackTag(request: AddFeedbackTagRequest): Promise<AddFeedbackTagResponse>;
  RemoveFeedbackTag(request: RemoveFeedbackTagRequest): Promise<RemoveFeedbackTagResponse>;
  BatchUpdateFeedbackTags(request: BatchUpdateFeedbackTagsRequest): Promise<BatchUpdateFeedbackTagsResponse>;
}